refactor: remove IPC from mouse_util (#38183)
